Feat: donate image similiar check and safe close command

+++ src/core/graphics/message.js
@@ -7,8 +7,10 @@ module.exports = {
      * @param {string} content message content
      * @param {number} deleteAfter delete delay (ms)
      */
-    async send(channel, content, deleteAfter) {
+    async send(channel, content, deleteAfter = 0) {
         const message = await channel.send(content);
+        if (deleteAfter < 0) throw new Error("only accept number > 0");
+        else if (deleteAfter === 0) return;
         setTimeout(async () => await message.delete(), deleteAfter);
     },
 };
